+++ MiddleProject/src/main/java/com/yedam/common/FrontController.java
@@ -36,6 +36,7 @@ import com.yedam.user.control.DeleteUserControl;
 import com.yedam.user.control.LoginControl;
 import com.yedam.user.control.LoginFormControl;
 import com.yedam.user.control.LogoutControl;
+import com.yedam.user.control.ModifyUserControl;
 import com.yedam.user.control.ModifyUserFormControl;
 import com.yedam.user.control.MypageControl;
 import com.yedam.user.control.MypageViewControl;
@@ -84,6 +85,7 @@ public class FrontController extends HttpServlet{
 		map.put("/mypagePw.do", new MypageControl());
 		map.put("/mypageView.do", new MypageViewControl());
 		map.put("/deleteUser.do", new DeleteUserControl());
+		map.put("/modifyUser.do", new ModifyUserControl());
 		map.put("/modifyUserForm.do", new ModifyUserFormControl());
 
 		//note
