+++ MiddleProject/src/main/java/com/yedam/user/mapper/UserMapper.xml
@@ -36,7 +36,7 @@
 <!-- 회원정보 수정 -->
 <update id="modifyUser" parameterType="com.yedam.user.domain.UserVO">
     UPDATE users
-    SET user_pw=#{userPw}, user_phone=#{userPhone}, useradd=#{userAdd}, user_nm=#{userNm}
+    SET user_pw=#{userPw}, user_phone=#{userPhone}, user_add=#{userAdd}
     WHERE user_id=#{userId}
 </update>
