+++ MiddleProject/src/main/webapp/.gitignore
@@ -1,0 +1,1 @@
+/META-INF/
