+++ MiddleProject/src/main/webapp/META-INF/MANIFEST.MF
@@ -1,3 +1,4 @@
 Manifest-Version: 1.0
-Class-Path: 
+Build-Jdk-Spec: 17
+Created-By: Maven Integration for Eclipse
